portolio/requirements.txt
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==4.1.0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
click==7.1.2
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
Flask==1.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imbalanced-learn==0.8.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
nltk==3.6.7
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.13.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyrsistent==0.18.0
python-dateutil==2.8.2
QtPy==2.0.1
regex==2022.10.31
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.0.1
tenacity==8.2.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
textblob==0.17.1
threadpoolctl==3.1.0
torch==1.9.0
torchaudio==0.7.2
torchvision==0.8.2+cpu
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.2
wordcloud==1.8.2.2
wrapt==1.12.1
zipp==3.6.0
